Update and add tests for trace collector, fix some bugs.
strategy.py: fix some bugs found during the testing
systemtap.py: bugfixes and small changes
systemtap_script.py: minor bugfixes
test_cli.py: move the trace collector tests to test_collect.py
test_collect.py: update and add tests
tests/collect_trace/: update the comparison outputs
