+++ tests/collect_trace/cmp_script.txt
@@ -1,14 +1,14 @@
-global samp_0[3]
+global samp_0[2]
 probe process("cmp").begin {
 	samp_0[0] = 1
-	samp_0[1] = 2
-	samp_0[2] = 3
+	samp_0[1] = 3
+	printf("cmp")
 }
 
 probe process("cmp").function("main").call? {
 	samp_0[0] ++
 	if(samp_0[0] == 2) {
-		printf("0 %s main\n", thread_indent(1))
+		printf("0 %smain\n", thread_indent(1))
 		samp_0[0] = 0
 	}
 }
@@ -19,7 +19,7 @@ probe process("cmp").function("main").return? {
 }
 
 probe process("cmp").function("test").call? {
-	printf("0 %s test\n", thread_indent(1))
+	printf("0 %stest\n", thread_indent(1))
 }
 
 probe process("cmp").function("test").return? {
@@ -27,23 +27,18 @@ probe process("cmp").function("test").return? {
 }
 
 probe process("cmp").mark("BEFORE_CYCLE") {
-	samp_0[1] ++
-	if(samp_0[1] == 3) {
-		printf("3 %s BEFORE_CYCLE\n", thread_indent(0))
-		samp_0[1] = 0
-	}
+	printf("3 %sBEFORE_CYCLE\n", thread_indent(0))
 }
+
 probe process("cmp").mark("BEFORE_CYCLE_end") {
-	if(samp_0[1] == 0) {
-		printf("4 %s BEFORE_CYCLE_end\n", thread_indent(0))
-	}
+	printf("4 %sBEFORE_CYCLE_end\n", thread_indent(0))
 }
 
 probe process("cmp").mark("TEST_SINGLE") {
-	samp_0[2] ++
-	if(samp_0[2] == 4) {
+	samp_0[1] ++
+	if(samp_0[1] == 4) {
 		printf("2 %s TEST_SINGLE\n", thread_indent(0))
-		samp_0[2] = 0
+		samp_0[1] = 0
 	}
 }
 probe process("cmp").mark("TEST_SINGLE2") {
@@ -51,5 +46,5 @@ probe process("cmp").mark("TEST_SINGLE2") {
 }
 
 probe process("cmp").end {
-	printf("end\n")
+	printf("cmp")
 }
