+++ tests/collect_trace/strategy5_script.txt
@@ -11,68 +11,69 @@ probe process("cmp").begin {
 	samp_0[8] = 9
 	samp_0[9] = 9
 	samp_0[10] = 9
+	printf("cmp")
 }
 
-probe process("cmp").function("main").call? {
+probe process("cmp").function("_Z10BubbleSortPii").call? {
 	samp_0[0] ++
 	if(samp_0[0] == 10) {
-		printf("0 %s main\n", thread_indent(1))
+		printf("0 %s_Z10BubbleSortPii\n", thread_indent(1))
 		samp_0[0] = 0
 	}
 }
-probe process("cmp").function("main").return? {
+probe process("cmp").function("_Z10BubbleSortPii").return? {
 	if(samp_0[0] == 0) {
 		printf("1 %s\n", thread_indent(-1))
 	}
 }
 
-probe process("cmp").function("_Z10BubbleSortPii").call? {
+probe process("cmp").function("_Z10InsertSortPii").call? {
 	samp_0[1] ++
 	if(samp_0[1] == 10) {
-		printf("0 %s _Z10BubbleSortPii\n", thread_indent(1))
+		printf("0 %s_Z10InsertSortPii\n", thread_indent(1))
 		samp_0[1] = 0
 	}
 }
-probe process("cmp").function("_Z10BubbleSortPii").return? {
+probe process("cmp").function("_Z10InsertSortPii").return? {
 	if(samp_0[1] == 0) {
 		printf("1 %s\n", thread_indent(-1))
 	}
 }
 
-probe process("cmp").function("_Z10InsertSortPii").call? {
+probe process("cmp").function("_Z12BadPartitionPiii").call? {
 	samp_0[2] ++
 	if(samp_0[2] == 10) {
-		printf("0 %s _Z10InsertSortPii\n", thread_indent(1))
+		printf("0 %s_Z12BadPartitionPiii\n", thread_indent(1))
 		samp_0[2] = 0
 	}
 }
-probe process("cmp").function("_Z10InsertSortPii").return? {
+probe process("cmp").function("_Z12BadPartitionPiii").return? {
 	if(samp_0[2] == 0) {
 		printf("1 %s\n", thread_indent(-1))
 	}
 }
 
-probe process("cmp").function("_Z12BadPartitionPiii").call? {
+probe process("cmp").function("_Z12QuickSortBadPii").call? {
 	samp_0[3] ++
 	if(samp_0[3] == 10) {
-		printf("0 %s _Z12BadPartitionPiii\n", thread_indent(1))
+		printf("0 %s_Z12QuickSortBadPii\n", thread_indent(1))
 		samp_0[3] = 0
 	}
 }
-probe process("cmp").function("_Z12BadPartitionPiii").return? {
+probe process("cmp").function("_Z12QuickSortBadPii").return? {
 	if(samp_0[3] == 0) {
 		printf("1 %s\n", thread_indent(-1))
 	}
 }
 
-probe process("cmp").function("_Z12QuickSortBadPii").call? {
+probe process("cmp").function("_Z4SwapRiS_").call? {
 	samp_0[4] ++
 	if(samp_0[4] == 10) {
-		printf("0 %s _Z12QuickSortBadPii\n", thread_indent(1))
+		printf("0 %s_Z4SwapRiS_\n", thread_indent(1))
 		samp_0[4] = 0
 	}
 }
-probe process("cmp").function("_Z12QuickSortBadPii").return? {
+probe process("cmp").function("_Z4SwapRiS_").return? {
 	if(samp_0[4] == 0) {
 		printf("1 %s\n", thread_indent(-1))
 	}
@@ -81,7 +82,7 @@ probe process("cmp").function("_Z12QuickSortBadPii").return? {
 probe process("cmp").function("_Z4swapPiii").call? {
 	samp_0[5] ++
 	if(samp_0[5] == 10) {
-		printf("0 %s _Z4swapPiii\n", thread_indent(1))
+		printf("0 %s_Z4swapPiii\n", thread_indent(1))
 		samp_0[5] = 0
 	}
 }
@@ -91,71 +92,71 @@ probe process("cmp").function("_Z4swapPiii").return? {
 	}
 }
 
-probe process("cmp").function("_Z4SwapRiS_").call? {
+probe process("cmp").function("_Z8HeapSortPii").call? {
 	samp_0[6] ++
 	if(samp_0[6] == 10) {
-		printf("0 %s _Z4SwapRiS_\n", thread_indent(1))
+		printf("0 %s_Z8HeapSortPii\n", thread_indent(1))
 		samp_0[6] = 0
 	}
 }
-probe process("cmp").function("_Z4SwapRiS_").return? {
+probe process("cmp").function("_Z8HeapSortPii").return? {
 	if(samp_0[6] == 0) {
 		printf("1 %s\n", thread_indent(-1))
 	}
 }
 
-probe process("cmp").function("_Z8HeapSortPii").call? {
+probe process("cmp").function("_Z9PartitionPiii").call? {
 	samp_0[7] ++
 	if(samp_0[7] == 10) {
-		printf("0 %s _Z8HeapSortPii\n", thread_indent(1))
+		printf("0 %s_Z9PartitionPiii\n", thread_indent(1))
 		samp_0[7] = 0
 	}
 }
-probe process("cmp").function("_Z8HeapSortPii").return? {
+probe process("cmp").function("_Z9PartitionPiii").return? {
 	if(samp_0[7] == 0) {
 		printf("1 %s\n", thread_indent(-1))
 	}
 }
 
-probe process("cmp").function("_Z9PartitionPiii").call? {
+probe process("cmp").function("_Z9QuickSortPii").call? {
 	samp_0[8] ++
 	if(samp_0[8] == 10) {
-		printf("0 %s _Z9PartitionPiii\n", thread_indent(1))
+		printf("0 %s_Z9QuickSortPii\n", thread_indent(1))
 		samp_0[8] = 0
 	}
 }
-probe process("cmp").function("_Z9PartitionPiii").return? {
+probe process("cmp").function("_Z9QuickSortPii").return? {
 	if(samp_0[8] == 0) {
 		printf("1 %s\n", thread_indent(-1))
 	}
 }
 
-probe process("cmp").function("_Z9QuickSortPii").call? {
+probe process("cmp").function("_Z9repairTopPiii").call? {
 	samp_0[9] ++
 	if(samp_0[9] == 10) {
-		printf("0 %s _Z9QuickSortPii\n", thread_indent(1))
+		printf("0 %s_Z9repairTopPiii\n", thread_indent(1))
 		samp_0[9] = 0
 	}
 }
-probe process("cmp").function("_Z9QuickSortPii").return? {
+probe process("cmp").function("_Z9repairTopPiii").return? {
 	if(samp_0[9] == 0) {
 		printf("1 %s\n", thread_indent(-1))
 	}
 }
 
-probe process("cmp").function("_Z9repairTopPiii").call? {
+probe process("cmp").function("main").call? {
 	samp_0[10] ++
 	if(samp_0[10] == 10) {
-		printf("0 %s _Z9repairTopPiii\n", thread_indent(1))
+		printf("0 %smain\n", thread_indent(1))
 		samp_0[10] = 0
 	}
 }
-probe process("cmp").function("_Z9repairTopPiii").return? {
+probe process("cmp").function("main").return? {
 	if(samp_0[10] == 0) {
 		printf("1 %s\n", thread_indent(-1))
 	}
 }
 
 probe process("cmp").end {
-	printf("end\n")
+	printf("cmp")
 }
