+++ tests/collect_trace/strategy7_script.txt
@@ -1,13 +1,9 @@
-probe process("cmp").function("main").call? {
-	printf("0 %s main\n", thread_indent(1))
-}
-
-probe process("cmp").function("main").return? {
-	printf("1 %s\n", thread_indent(-1))
+probe process("cmp").begin {
+	printf("cmp")
 }
 
 probe process("cmp").function("_Z10BubbleSortPii").call? {
-	printf("0 %s _Z10BubbleSortPii\n", thread_indent(1))
+	printf("0 %s_Z10BubbleSortPii\n", thread_indent(1))
 }
 
 probe process("cmp").function("_Z10BubbleSortPii").return? {
@@ -15,7 +11,7 @@ probe process("cmp").function("_Z10BubbleSortPii").return? {
 }
 
 probe process("cmp").function("_Z10InsertSortPii").call? {
-	printf("0 %s _Z10InsertSortPii\n", thread_indent(1))
+	printf("0 %s_Z10InsertSortPii\n", thread_indent(1))
 }
 
 probe process("cmp").function("_Z10InsertSortPii").return? {
@@ -23,7 +19,7 @@ probe process("cmp").function("_Z10InsertSortPii").return? {
 }
 
 probe process("cmp").function("_Z12BadPartitionPiii").call? {
-	printf("0 %s _Z12BadPartitionPiii\n", thread_indent(1))
+	printf("0 %s_Z12BadPartitionPiii\n", thread_indent(1))
 }
 
 probe process("cmp").function("_Z12BadPartitionPiii").return? {
@@ -31,31 +27,31 @@ probe process("cmp").function("_Z12BadPartitionPiii").return? {
 }
 
 probe process("cmp").function("_Z12QuickSortBadPii").call? {
-	printf("0 %s _Z12QuickSortBadPii\n", thread_indent(1))
+	printf("0 %s_Z12QuickSortBadPii\n", thread_indent(1))
 }
 
 probe process("cmp").function("_Z12QuickSortBadPii").return? {
 	printf("1 %s\n", thread_indent(-1))
 }
 
-probe process("cmp").function("_Z4swapPiii").call? {
-	printf("0 %s _Z4swapPiii\n", thread_indent(1))
-}
-
-probe process("cmp").function("_Z4swapPiii").return? {
-	printf("1 %s\n", thread_indent(-1))
-}
-
 probe process("cmp").function("_Z4SwapRiS_").call? {
-	printf("0 %s _Z4SwapRiS_\n", thread_indent(1))
+	printf("0 %s_Z4SwapRiS_\n", thread_indent(1))
 }
 
 probe process("cmp").function("_Z4SwapRiS_").return? {
 	printf("1 %s\n", thread_indent(-1))
 }
 
+probe process("cmp").function("_Z4swapPiii").call? {
+	printf("0 %s_Z4swapPiii\n", thread_indent(1))
+}
+
+probe process("cmp").function("_Z4swapPiii").return? {
+	printf("1 %s\n", thread_indent(-1))
+}
+
 probe process("cmp").function("_Z8HeapSortPii").call? {
-	printf("0 %s _Z8HeapSortPii\n", thread_indent(1))
+	printf("0 %s_Z8HeapSortPii\n", thread_indent(1))
 }
 
 probe process("cmp").function("_Z8HeapSortPii").return? {
@@ -63,7 +59,7 @@ probe process("cmp").function("_Z8HeapSortPii").return? {
 }
 
 probe process("cmp").function("_Z9PartitionPiii").call? {
-	printf("0 %s _Z9PartitionPiii\n", thread_indent(1))
+	printf("0 %s_Z9PartitionPiii\n", thread_indent(1))
 }
 
 probe process("cmp").function("_Z9PartitionPiii").return? {
@@ -71,7 +67,7 @@ probe process("cmp").function("_Z9PartitionPiii").return? {
 }
 
 probe process("cmp").function("_Z9QuickSortPii").call? {
-	printf("0 %s _Z9QuickSortPii\n", thread_indent(1))
+	printf("0 %s_Z9QuickSortPii\n", thread_indent(1))
 }
 
 probe process("cmp").function("_Z9QuickSortPii").return? {
@@ -79,13 +75,21 @@ probe process("cmp").function("_Z9QuickSortPii").return? {
 }
 
 probe process("cmp").function("_Z9repairTopPiii").call? {
-	printf("0 %s _Z9repairTopPiii\n", thread_indent(1))
+	printf("0 %s_Z9repairTopPiii\n", thread_indent(1))
 }
 
 probe process("cmp").function("_Z9repairTopPiii").return? {
 	printf("1 %s\n", thread_indent(-1))
 }
 
+probe process("cmp").function("main").call? {
+	printf("0 %smain\n", thread_indent(1))
+}
+
+probe process("cmp").function("main").return? {
+	printf("1 %s\n", thread_indent(-1))
+}
+
 probe process("cmp").end {
-	printf("end\n")
+	printf("cmp")
 }
